マイグレート周りのテストを更に整理した
